Update document DPI to match CSS standard.
This is a change applied automatically upon opening a document using Inkscape
0.92 or later. Explanation here: https://inkscape.org/en/learn/faq/#dpi_change

---

Why did document DPI change in Inkscape 0.92?

To prepare objects whose size is given in physical units like inches (in),
millimeters (mm) or point (pt) for rendering on your computer screen, Inkscape
needs to know the conversion factor from these "real world" units to the purely
digital size of a pixel (px).

Earlier versions of SVG and CSS specifications did not define a fixed conversion
between those units. It was basically up to the application to determine
something useful and - short of requiring users to measure the actual physical
size of one pixel on their monitor - a common approach was to define a "pixel
density" of 90 dots per inch (DPI), i.e. a pixel is fixed to be 1/90 inch in
size. Inkscape 0.91 and earlier followed this approach and used a fixed value of
90 DPI for the conversion.

Recent versions of the CSS specification changed this detail as it turned out to
be impractical: conversion factors varied widely across different systems and
applications and they often used "arbitrary" fixed conversion factors of their
own choice. To unify behavior, a pixel density of 96 DPI is now the universal
standard which is followed by all web browsers and is also part of the upcoming
SVG2 standard.

In order to match the rendering of browsers and become compatible to SVG2,
Inkscape has adjusted the pixel density used internally from 90 DPI to 96 DPI.
While this ensures that your artwork will always look the same, regardless how
or where it's shown, it also means that old documents have to be scaled if their
size should stay identical:

- a document that is 90 pixels wide had a physical width of exactly 1 inch in
  Inkscape 0.91 (90 px / 90 DPI = 1 in)

- the same document will be 6% smaller in Inkscape 0.92 (90 px / 96 DPI = 0.9375 in)