Change SVGs to load new svg-styles.css, which defines Open Sans.
The SVGs will display correctly even without the style sheet if the Open Sans
font is installed locally. When Open Sans is not installed locally, there's no
way to display the SVG correctly without referring to external resources
somehow. Linking to a shared CSS file seems like as good an option as any, and
it's convenient.